Logger

An R6 class for flexible logging with customizable output.

Methods

Public methods

Logger$new()

Logger$set_level()

Logger$error()

Logger$warn()

Logger$info()

Logger$clone()

Method new()

Create a new Logger object.

Usage

Logger$new(
  level = LogLevel$INFO,
  file_path = NULL,
  print_fn = function(x) cat(x, "\n")
)

Arguments

level

The minimum log level to output. Default is LogLevel$INFO.

file_path

Character; the path to a file to save log entries to. Default is NULL.

print_fn

Function; custom print function to use for console output. Should accept a single character string as input. Default uses cat with a newline.

Returns

A new `Logger` object.

Examples

logger <- Logger$new(level = LogLevel$WARNING, file_path = "log.txt")

Method set_level()

Set the minimum log level.

Usage

Logger$set_level(level)

Arguments

level

The new minimum log level to set.

Examples

logger <- Logger$new()
logger$set_level(LogLevel$WARNING)

Method error()

Log an error message.

Usage

Logger$error(msg, data = NULL, error = NULL)

Arguments

msg

Character; the error message to log.

data

Optional; additional data to include in the log entry.

error

Optional; an error object to include in the log entry.

Examples

logger <- Logger$new()
logger$error("An error occurred", data = list(x = 1), error = simpleError("Oops!"))

Method warn()

Log a warning message.

Usage

Logger$warn(msg, data = NULL)

Arguments

msg

Character; the warning message to log.

data

Optional; additional data to include in the log entry.

Examples

logger <- Logger$new()
logger$warn("This is a warning", data = list(reason = "example"))

Method info()

Log an info message.

Usage

Logger$info(msg, data = NULL)

Arguments

msg

Character; the info message to log.

data

Optional; additional data to include in the log entry.

Examples

logger <- Logger$new()
logger$info("Operation completed successfully", data = list(duration = 5.2))

Method clone()

The objects of this class are cloneable with this method.

Usage

Logger$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create a basic logger
logger <- Logger$new()
logger$info("This is an info message")
#> 2024-08-03T15:49:00.628Z INFO    This is an info message 
logger$warn("This is a warning")
#> 2024-08-03T15:49:00.632Z WARNING This is a warning 
logger$error("This is an error")
#> 2024-08-03T15:49:00.642Z ERROR   This is an error 

# Create a logger with custom settings
custom_logger <- Logger$new(
  level = LogLevel$WARNING,
  file_path = tempfile("log_"),
  print_fn = message
)
custom_logger$info("This won't be logged")
custom_logger$warn("This will be logged")
#> 2024-08-03T15:49:00.647Z WARNING This will be logged

# Change log level
custom_logger$set_level(LogLevel$INFO)
custom_logger$info("Now this will be logged")
#> 2024-08-03T15:49:00.649Z INFO    Now this will be logged

## ------------------------------------------------
## Method `Logger$new`
## ------------------------------------------------

logger <- Logger$new(level = LogLevel$WARNING, file_path = "log.txt")

## ------------------------------------------------
## Method `Logger$set_level`
## ------------------------------------------------

logger <- Logger$new()
logger$set_level(LogLevel$WARNING)

## ------------------------------------------------
## Method `Logger$error`
## ------------------------------------------------

logger <- Logger$new()
logger$error("An error occurred", data = list(x = 1), error = simpleError("Oops!"))
#> 2024-08-03T15:49:00.655Z ERROR   An error occurred
#> Data:
#> {
#>   "x": 1
#> }
#> Error:
#> {
#>   "name": "simpleError",
#>   "message": "Oops!",
#>   "call": "NULL"
#> } 

## ------------------------------------------------
## Method `Logger$warn`
## ------------------------------------------------

logger <- Logger$new()
logger$warn("This is a warning", data = list(reason = "example"))
#> 2024-08-03T15:49:00.658Z WARNING This is a warning
#> Data:
#> {
#>   "reason": "example"
#> } 

## ------------------------------------------------
## Method `Logger$info`
## ------------------------------------------------

logger <- Logger$new()
logger$info("Operation completed successfully", data = list(duration = 5.2))
#> 2024-08-03T15:49:00.660Z INFO    Operation completed successfully
#> Data:
#> {
#>   "duration": 5.2
#> }