Append log messages to a file and stdout as well

Append log messages to a file and stdout as well

appender_tee(file, append = TRUE)

Arguments

file
path

append
boolean passed to cat defining if the file should be overwritten with the most recent log message instead of appending

Value

function taking lines argument

See also