Format a log message with glue

By default, this layout includes the log level of the log record as per log_levels, the current timestamp and the actual log message -- that you can override via calling layout_glue_generator directly. For colorized output, see layout_glue_colors.

layout_glue(level, msg, namespace = NA_character_,
  .logcall = sys.call(), .topcall = sys.call(-1),
  .topenv = parent.frame())

Arguments

level
log level, see log_levels for more details

msg
string message

namespace
string referring to the logger environment / config to be used to override the target of the message record to be used instead of the default namespace, which is defined by the R package name from which the logger was called, and falls back to a common, global namespace.

.logcall
the logging call being evaluated (useful in formatters and layouts when you want to have access to the raw, unevaluated R expression)

.topcall
R expression from which the logging function was called (useful in formatters and layouts to extract the calling function's name or arguments)

.topenv
original frame of the .topcall calling function where the formatter function will be evaluated and that is used to look up the namespace as well via logger:::top_env_name

Value

character vector

See also