Generate log layout function using common variables available via glue syntax

format is passed to glue with access to the below variables:

msg: the actual log message

further variables set by get_logger_meta_variables

layout_glue_generator(format = "{level} [{format(time, \"%Y-%d-%m %H:%M:%S\")}] {msg}")

Arguments

format
glue-flavored layout of the message using the above variables

Value

function taking level and msg arguments - keeping the original call creating the generator in the generator attribute that is returned when calling log_layout for the currently used layout

See also

Examples

# NOT RUN {
example_layout <- layout_glue_generator(
  format = '{node}/{pid}/{ns}/{ans}/{topenv}/{fn} {time} {level}: {msg}')
example_layout(INFO, 'try {runif(1)}')

log_layout(example_layout)
log_info('try {runif(1)}')
# }