Adds the skip_formatter attribute to an object so that logger will skip calling the formatter function on the object(s) to be logged

Adds the skip_formatter attribute to an object so that logger will skip calling the formatter function on the object(s) to be logged

skip_formatter(message, ...)

Arguments

message
character vector directly passed to the appender function in logger

...
should be never set

Value

character vector with skip_formatter attribute set to TRUE