Generate logging function

Available variables to be used in the msg_format provided by logger:::get_logger_meta_variables:

msg: the actual log message

level: log level, eg INFO

time: current time formatted as time_format

namespace: R package (if any) calling the logging function

call: parent call (if any) calling the logging function

fn: function's (if any) name calling the logging function

user: name of the real user id as reported by Sys.info

pid: the process identification number of the R session

node: name by which the machine is known on the network as reported by Sys.info

layout_glue_generator(format = "{level} [{format(time, \"%Y-%d-%m %H:%M:%S\")}] {msg}")

Arguments

format
glue-flavored layout of the message using the above variables

Value

function taking level and msg arguments - keeping the original call creating the generator in the generator attribute that is returned when calling log_layout() for the currently used layout

Examples

# NOT RUN {
logger <- layout_glue_generator(format = '{node}/{pid}/{namespace}/{fn} {time} {level}: {msg}')
logger(FATAL, 'try {runif(1)}')

log_layout(logger)
log_info('try {runif(1)}')
# }