Generate log layout function using common variables available via glue syntax

format is passed to glue with access to the below variables:

msg: the actual log message

further variables set by get_logger_meta_variables()

Usage

layout_glue_generator(
  format = "{level} [{format(time, \"%Y-%m-%d %H:%M:%S\")}] {msg}"
)

Arguments

format

glue-flavored layout of the message using the above variables

Value

function taking level and msg arguments - keeping the original call creating the generator in the generator attribute that is returned when calling log_layout() for the currently used layout

See also

Examples

example_layout <- layout_glue_generator(
  format = "{node}/{pid}/{ns}/{ans}/{topenv}/{fn} {time} {level}: {msg}"
)
example_layout(INFO, "try {runif(1)}")
#> fv-az1776-477/7978/NA/global/R_GlobalEnv/eval 2024-10-20 19:57:50.250538 INFO: try {runif(1)}

log_layout(example_layout)
log_info("try {runif(1)}")
#> fv-az1776-477/7978/global/global/R_GlobalEnv/eval 2024-10-20 19:57:50.253134 INFO: try 0.0807501375675201