Logs a message in a very visible way

Logs a message in a very visible way

Usage

log_with_separator(
  ...,
  level = INFO,
  namespace = NA_character_,
  separator = "=",
  width = 80
)

Arguments

...

R objects that can be converted to a character vector via the active message formatter function

level

log level, see log_levels() for more details

namespace

string referring to the logger environment / config to be used to override the target of the message record to be used instead of the default namespace, which is defined by the R package name from which the logger was called, and falls back to a common, global namespace.

separator

character to be used as a separator

width

max width of message – longer text will be wrapped into multiple lines

See also

Examples

log_with_separator("An important message")
#> ================================================================================
#> = An important message                                                         =
#> ================================================================================
log_with_separator("Some critical KPI down!!!", separator = "$")
#> $$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$
#> $ Some critical KPI down!!!                                                    $
#> $$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$
log_with_separator("This message is worth a {1e3} words")
#> ================================================================================
#> = This message is worth a 1000 words                                           =
#> ================================================================================
log_with_separator(paste(
  "A very important message with a bunch of extra words that will",
  "eventually wrap into a multi-line message for our quite nice demo :wow:"
))
#> ================================================================================
#> = A very important message with a bunch of extra words that will eventually    =
#> = wrap into a multi-line message for our quite nice demo :wow:                 =
#> ================================================================================
log_with_separator(
  paste(
    "A very important message with a bunch of extra words that will",
    "eventually wrap into a multi-line message for our quite nice demo :wow:"
  ),
  width = 60
)
#> ============================================================
#> = A very important message with a bunch of extra words     =
#> = that will eventually wrap into a multi-line message for  =
#> = our quite nice demo :wow:                                =
#> ============================================================
log_with_separator("Boo!", level = FATAL)
#> ================================================================================
#> = Boo!                                                                         =
#> ================================================================================
log_layout(layout_blank)
log_with_separator("Boo!", level = FATAL)
#> ================================================================================
#> = Boo!                                                                         =
#> ================================================================================
logger <- layout_glue_generator(format = "{node}/{pid}/{namespace}/{fn} {time} {level}: {msg}")
log_layout(logger)
log_with_separator("Boo!", level = FATAL, width = 120)
#> fv-az1249-914/6185/global/eval 2024-08-27 20:42:16.725876 FATAL: =======================================================
#> fv-az1249-914/6185/global/log_with_separator 2024-08-27 20:42:16.72671 FATAL: = Boo!                                                =
#> fv-az1249-914/6185/global/eval 2024-08-27 20:42:16.728133 FATAL: =======================================================